swagger: add flag to add title
Currently the title of the generated swagger file is the name of the
file. This provides no user input to customize.

Let's add a new flag `-swagger_title` which will allow the user to
provide a title and this is propagated to the lower layers and used
during the swagger generation.

We still default to filename if no title is provided.

Signed-off-by: Karthik Nayak <karthik.188@gmail.com>